+++ src/main/java/com/diegoygabriela/backend_novalink/repository/DetalleRegaloRepository.java
@@ -16,18 +16,18 @@ public interface DetalleRegaloRepository extends JpaRepository<DetalleRegalo, Lo
     List<DetalleRegalo> findByParejaIdOrderByFechaDesc(Long parejaId);
     
     // Buscar regalos dados por un usuario específico
-    List<DetalleRegalo> findByRemitenteIdUsuarioOrderByFechaDesc(Integer remitenteId);
+    List<DetalleRegalo> findByRemitenteIdUsuarioOrderByFechaDesc(Long remitenteId);
     
     // Buscar regalos recibidos por un usuario específico
-    List<DetalleRegalo> findByReceptorIdUsuarioOrderByFechaDesc(Integer receptorId);
+    List<DetalleRegalo> findByReceptorIdUsuarioOrderByFechaDesc(Long receptorId);
     
     // Buscar regalos entre dos usuarios específicos
     @Query("SELECT dr FROM DetalleRegalo dr WHERE " +
            "(dr.remitente.idUsuario = :usuario1Id AND dr.receptor.idUsuario = :usuario2Id) OR " +
            "(dr.remitente.idUsuario = :usuario2Id AND dr.receptor.idUsuario = :usuario1Id) " +
            "ORDER BY dr.fecha DESC")
-    List<DetalleRegalo> findRegalosBetweenUsers(@Param("usuario1Id") Integer usuario1Id, 
-                                               @Param("usuario2Id") Integer usuario2Id);
+    List<DetalleRegalo> findRegalosBetweenUsers(@Param("usuario1Id") Long usuario1Id, 
+                                               @Param("usuario2Id") Long usuario2Id);
     
     // Buscar regalos por rango de fechas
     List<DetalleRegalo> findByFechaBetweenOrderByFechaDesc(LocalDateTime fechaInicio, LocalDateTime fechaFin);
